refactor(tools): move util yt-dlp and gallery-dl classes to a common util package
